# Fix performance regression in `Greedy`